refactor&features: Implemented a new controller method 'deleteFlower' to handle flower deletion by ID. Added exception handling to the 'deleteFlower' method to catch ProductNotFoundException and FlowerDeleteException. Returned appropriate HTTP responses with error messages when exceptions occur during flower deletion. Updated the 'FlowerServiceImpl' class to throw FlowerDeleteException when flower deletion encounters an unexpected error. Improved error handling in other controller methods such as 'getFlowerId', 'createFlower', and 'updateFlower' to provide more informative error messages to the front end.Created an Image controller and service where you can pull out a link to the picture. .

+++ server/src/main/java/com/example/flowersproject/repository/FlowerRepository.java
@@ -5,4 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FlowerRepository extends JpaRepository<FlowerEntity, Integer> {}
+public interface FlowerRepository extends JpaRepository<FlowerEntity, Long> {
+
+
+
+}
